Implemented challenge and accept/deny functionality.
In lobby, sending a challenge post request causes the server to emit a socket message to the target client room.
The target client can then either send an accept or deny post request. If accept, the server sends an accept message to the challenger.

Upon accept, each client redirects to the game. This is done in accept message handler in challengee, and in click handler in challenger.